# frontend/src/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BotDashboard API",
    "description": "API per la gestione del Bot Dashboard con telemetria, comandi e statistiche real-time",
    "version": "v1"
  },
  "paths": {
...
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets all users (Admin only).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIEnumerableApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Creates a new user (Admin only).",
        "parameters": [
          {
            "name": "roleName",
            "in": "query",
            "description": "Optional role name (Admin, User, BotOperator). Default is based on IsAdmin flag.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The create user request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets a user by ID (Authenticated users can see their own profile, Admin can see all).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Deletes a user (Admin only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/roles-and-permissions": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets all roles and permissions for a user (Admin only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesAndPermissionsResponseApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/roles": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Assigns a role to a user (Admin only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The role assignment request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/roles/{roleName}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Removes a role from a user (Admin only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleName",
            "in": "path",
            "description": "The role name to remove.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/permissions": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Assigns a permission to a user (Admin only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The permission assignment request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/permissions/{permission}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Removes a permission from a user (Admin only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "permission",
            "in": "path",
            "description": "The permission to remove.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignPermissionRequest": {
        "required": [
          "permission"
        ],
        "type": "object",
        "properties": {
          "permission": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AssignRoleRequest": {
        "required": [
          "roleName"
        ],
        "type": "object",
        "properties": {
          "roleName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CreateUserRequest": {
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ObjectApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Gets or sets a value indicating whether the operation was successful."
          },
          "message": {
            "type": "string",
            "description": "Gets or sets the message describing the result.",
            "nullable": true
          },
          "data": {
            "description": "Gets or sets the data payload.",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or sets the list of errors, if any.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Gets or sets the timestamp of the response.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Represents a standardized API response."
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "deprecated": true
          },
          "admin": {
            "type": "boolean"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Gets or sets a value indicating whether the operation was successful."
          },
          "message": {
            "type": "string",
            "description": "Gets or sets the message describing the result.",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/User"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or sets the list of errors, if any.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Gets or sets the timestamp of the response.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Represents a standardized API response."
      },
      "UserIEnumerableApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Gets or sets a value indicating whether the operation was successful."
          },
          "message": {
            "type": "string",
            "description": "Gets or sets the message describing the result.",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "description": "Gets or sets the data payload.",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or sets the list of errors, if any.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Gets or sets the timestamp of the response.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Represents a standardized API response."
      },
      "UserRolesAndPermissionsResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UserRolesAndPermissionsResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Gets or sets a value indicating whether the operation was successful."
          },
          "message": {
            "type": "string",
            "description": "Gets or sets the message describing the result.",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/UserRolesAndPermissionsResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or sets the list of errors, if any.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Gets or sets the timestamp of the response.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Represents a standardized API response."
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}